backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
colorama==0.4.6
fastapi==0.115.12
greenlet==3.1.1
h11==0.14.0
idna==3.10
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.21.1
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
typing_extensions==4.12.2
uvicorn==0.34.0
